Add a starter project for the Checkout kata.
This is a more direct translation of the Ruby example than we have been
working to so far.

I have not converted the RULES parameter of Checkout() for two reasons:
  1. The appropriate data structure will be a first-class collection, not
  an instance of a CLR collection.
  2. I didn't want to give too much of the game away up front!

Note that I have deliberately left *in*:
  - the Checkout.csproj.user file: this contains the command-line
    arguments for the debug instance.
  - the NCrunch files: just in case we get a candidate who wants to use
    this tool...

There is precisely one failing test at this point: GetTotal() must return
0 when no items are scanned.

NUnit is pulled in via NuGet.  No other libraries are included.